修正0430type錯誤導致專案壞掉無法進首頁
修改 cart 內的 userid type 為 string
DB 內修改 cart schema

修正跳轉不存在的 index.jsp 導致錯誤 在跳轉到不存在的 404.jsp

加入購物車功能仍然無法使用

+++ src/main/java/com/example/dao/CartDao.java
@@ -8,6 +8,6 @@ public interface CartDao {
 
     void saveCart(Cart cart);
     void saveCartDetail(CartDetail detail);
-    CartDetail getCartDetail(Integer cartId, Integer productId);
+    CartDetail getCartDetail(Integer cartId, long productId);
     void updateCartDetail(CartDetail detail);
 }
