+++ src/main/java/com/example/service/impl/ProductServiceImpl.java
@@ -28,7 +28,8 @@ public class ProductServiceImpl implements ProductService {
     }
     
     @Override
-    public Product getProductById(Integer id) {
+    public Product getProductById(long id) {
         return productDao.getProductById(id);
     }
+
 }
